Don't load always sdp yaml specification
Some tests dont use it so it's not needed to load it always
